次のアウトプット政策のため、ファイル整理
